Avoid pointless use of one-celled array (which also confounds Metrowerks C).
Pass the address of fetch_refspec_str directly to parse_refspec_internal()
instead of copying fetch_refspec_str to an array and passing that.

Metrowerks C won't initialize aggregates with values determined at runtime, and
we can't use the C++ translator for parse_refspec_internal() because it uses the
same identifier (refspec) for a variable as an existing struct type.

That, and there's no purpose to the array in the first place.