[FIX] Include <ctype.h> early so git's macros don't hose it
Include <ctype.h> *before* defining macros of its functions, in case it's
included later (which it is, in the case of fnmatch and xdiff).

The reason this has only become an issue now is that sys/poll.h was erroneously
including <ctype.h> instead of <sys/cdefs.h>, which was fixed to enable poll().