Exit without global destruction in start_command()
The environ storage is shared between parent and child of reexec() -- so
it's critical not to destroy it in the child.  This fixes a crash due to
memory corruption when using a git alias in MacRelix that only showed up
after modifying the environ storage to refrain from throwing exceptions.