Avoid broken PPC optimization of collect_parents()
Added in b5d887f90692515a6e9c99e7683533d12df3ade2, the collect_parents()
function is miscompiled by Metrowerks C++ 2.4.1 at an optimization level
of 2 or above.  The optimizer fails to recognize that remoteheads (which
is on the stack) can be changed by the call to commit_list_insert(), due
to its address being stored in remotes (which is passed as an argument),
and returns NULL (the initial value of remoteheads) instead of the value
it actually has by that point.  This led to unchecked NULL dereferences.

Included in e78cbf8cbb61edfbdef5d33262f9b8cf02a3afca are patches for the
NULL dereferences, but the underlying problem persists, yielding another
issue:  A simple fast-forward merge always reports "Already up-to-date."